+++ DTOS/RouteHistoryForGetAllDTO.cs
@@ -4,8 +4,8 @@
     {
         public long VehicleId { get; set; }
         public long? VehicleNumber { get; set; }
-        public float? LastLatitude { get; set; }
-        public float? LastLongitude { get; set; }
+        public double? LastLatitude { get; set; }
+        public double? LastLongitude { get; set; }
         public int? VehicleDirection { get; set; }
         public char? Status { get; set; }
     }
